+++ src/jenn/post_processing/_metrics.py
@@ -7,11 +7,11 @@
 import numpy as np
 
 
-def r_square(y_pred: np.ndarray, y_true: np.ndarray) -> np.ndarray:
+def rsquare(y_pred: np.ndarray, y_true: np.ndarray) -> np.ndarray:
     """Compute R-square value for each output.
 
     :param y_pred: predicted values, array of shape (n_y, m)
-    :param y_true: actuial values, array of shape (n_y, m)
+    :param y_true: actual values, array of shape (n_y, m)
     :return: R-Squared values for each predicted reponse
     """
     axis = y_true.ndim - 1